API/requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==42.0.2
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
packaging==23.2
pillow==10.2.0
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
